feat(ios_sidebar): modify with feedback
