+++ xpeapp_ios/XpeApp/XpeApp/Data/Model/Login/UserPasswordEditReturnEnum.swift
@@ -1,0 +1,13 @@
+//
+//  UserPasswordEditReturnEnum.swift
+//  XpeApp
+//
+//  Created by Théo Lebègue on 14/01/2025.
+//
+
+
+public enum UserPasswordEditReturnEnum: String, Codable {
+    case success
+    case invalidInitialPassword
+    case newPasswordsNotMatch
+}
