--- xpeapp_ios/XpeApp/XpeApp/Untitled.swift
@@ -1,7 +1,0 @@
-//
-//  Untitled.swift
-//  XpeApp
-//
-//  Created by Théo Lebègue on 06/01/2025.
-//
-
